bytcr-rt5651: Change Speaker/Headphone en/disable sequences to fix switching
pulseaudio will run the DisableSequence of the current playback device
before running the EnableSequence of the new playback device.

This causes the Platform Clock and BIAS to temporarily get turned off which
breaks audio-streams which are playing when switching.

This commit moves the disabling to the EnableSequence of the other device
fixing this.

Signed-off-by: Hans de Goede <hdegoede@redhat.com>