chtnau8824: Fix 'THD for key voice command' value
According to:
https://source.android.com/devices/accessories/headset/jack-headset-spec
we should detect anything between 210 - 290 ohm as Function B.

I've 2 original Android headsets (shipped with Nexus devices) and both
have a value 215 ohm between the 2nd ring and the sleeve (according to
the spec this should be 240) when pressing Volume Up / Function B and
with the 'THD for key voice command' of 20 used before this commit this
gets recognized as Function D.

This commit changes the threshold value to 16, which according to my
measurements matches the 210 ohm from the spec, fixing the Volume Up
button not working. This has been tested with 2 official android headsets
on 2 different model tablets with a nau8824 codec.

Note I also have a clone headset which reports 170 ohm for Volume Up,
which requires changing the threshold value to 13 to work but that takes
the threshold well into the 110 - 180 ohm range reserved for Function D,
so that headset simply is just out of spec.

Signed-off-by: Hans de Goede <hdegoede@redhat.com>